Tier_2/06_data_engineering/final_project/requirements.txt
pyspark==3.5.5
kafka-python==2.1.2
apache-airflow==2.10.5
python-dotenv==1.1.0
pandas==2.2.3
chispa==0.11.1
pytest-mock==3.14.0
